% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eAnalytics-package.R
\docType{package}
\name{eAnalytics-package}
\alias{eAnalytics-package}
\title{eAnalytics: Dynamic Web-Based Analytics for the Energy Industry}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A 'Shiny' web application for energy industry analytics. Take an overview of the industry, measure Key Performance Indicators, identify changes in the industry over time, and discover new relationships in the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://paulgovan.github.io/eAnalytics/}
  \item \url{https://github.com/paulgovan/eAnalytics}
  \item Report bugs at \url{https://github.com/paulgovan/eAnalytics/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
