% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eAnalytics.R
\name{eAnalytics}
\alias{eAnalytics}
\title{Launch eAnalytics Shiny Application}
\usage{
eAnalytics()
}
\value{
The function does not return a value; it launches the Shiny app.
}
\description{
This function launches the eAnalytics Shiny application, which provides
interactive visualizations and analyses of energy data. The application
includes features such as time series plots, motion charts, data tables,
and geographic maps.
}
\examples{
if (interactive()) {
  eAnalytics()
}
}
\seealso{
\url{http://paulgovan.github.io/eAnalytics/}
}
