% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eAnalytics.R
\name{eAnalytics}
\alias{eAnalytics}
\title{Interactive and dynamic web analytics for the energy industry.}
\usage{
eAnalytics()
}
\description{
eAnalytics is a shiny web application for energy industry analytics.
eAnalytics has the largest open database of US energy industry information,
providing interactive and dynamic web analytics to industry stakeholders.
}
\examples{
if (interactive()) {
  eAnalytics()
}
}

